Fix #85 - new player doesn't break graph anymore
`oldest_game` was null when there was a new player being added because
there weren't any games in the system yet. Remedied this by doing a null
check as well as adding a "tracking since *today*" tag when it's a new
player.

Also fixed small typo in sharelink update, f=`something` was putting the
n value instead of the f value in.

Also removed old, commented out sharelink behaviour.